Introduce CM_Migration_Manager
